% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{compute_likelihood}
\alias{compute_likelihood}
\title{Compute likelihood for each additional simulation across timepoints}
\usage{
compute_likelihood(
  prevalence_map,
  simulated_prevalences,
  amis_params,
  likelihoods = NULL
)
}
\arguments{
\item{prevalence_map}{A list with one entry for each timepoint.
Each entry must be a list containing objects \code{data} (an L x M matrix of data);
and \code{likelihood} a function taking arguments \code{data} (a matrix of data as above),
\code{prevalence} (a matrix of output from the transmission model) and optional logical \code{log}, which returns the vector of (log)-likelihoods.
If a likelihood is not specified then it is assumed that
the data consist of samples from a geo-statistical model and empirical methods are used.}

\item{amis_params}{A list of parameters, e.g. from \code{\link{default_amis_params}}.}

\item{likelihoods}{An array with dimension n_tims,n_locs,n_sims -- ie timepoints x locations x simulations (optional).}

\item{An}{n x timepoints matrix of prevalences simulated from the transmission model.}
}
\value{
A larger array with the likelihoods of the new simulations joined to the existing array \code{likelihoods}.
}
\description{
Calls evaluate likelihood for each timepoint.
}
