% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{compute_prior_proposal_ratio}
\alias{compute_prior_proposal_ratio}
\title{Compute the prior/proposal ratio}
\usage{
compute_prior_proposal_ratio(components, param, prior_density, df, log)
}
\arguments{
\item{components}{A list of mixture components made of
\describe{
\item{\code{G}}{A numeric vector containing the number of components from each AMIS iteration}
\item{\code{Sigma}}{A list of covariace matrices from each component}
\item{\code{Mean}}{A list of means from each component}
\item{\code{probs}}{A list of probability weights for each component (unnormalised)}
}}

\item{param}{A matrix containing the sampled parameter vectors.}

\item{prior_density}{Vector containing the prior density of each sampled parameter vector.}

\item{df}{The degrees of freedom for the t-distributed proposal distribution.}

\item{log}{A logical indicating whether to work on the log scale.}
}
\value{
A vector containing the prior/proposal ratio for each row in
\code{param}
}
\description{
This function returns the ratio between the prior and proposal distribution
for each sampled parameter value (i.e. each row in \code{param}).
This function returns the first weight
See step (4) of the AMIS algorithm in
Integrating geostatistical maps and infectious disease transmission models
using adaptive multiple importance sampling.
Renata Retkute, Panayiota Touloupou, Maria-Gloria Basanez,
T. Deirdre Hollingsworth, Simon E.F. Spencer
Ann. Appl. Stat. 15 (4) 1980 - 1998, December 2021.
DOI: https://doi.org/10.1214/21-AOAS1486
}
