% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{compute_weight_matrix_histogram}
\alias{compute_weight_matrix_histogram}
\title{Compute weight matrix using empirical Radon-Nikodym derivative (with fixed breaks)}
\usage{
compute_weight_matrix_histogram(
  likelihoods,
  prev_sim,
  amis_params,
  weight_matrix
)
}
\arguments{
\item{likelihoods}{An n_sims x n_locs matrix of (log-)likelihoods
NB: transpose of slice of array.}

\item{prev_sim}{A vector containing the simulated prevalence value for each
parameter sample. (double)}

\item{amis_params}{A list of parameters, e.g. from \code{\link{default_amis_params}}}

\item{weight_matrix}{A matrix containing the current values of the weights.}
}
\value{
An updated weight matrix.
}
\description{
Compute matrix describing the weights for each parameter sampled, for each
location. One row per sample, one column per location.  Each weight
is computed based on the empirical Radon-Nikodym derivative, taking into account
geostatistical prevalence data for the specific location and the prevalence value
computed from the transmission model for the specific parameter sample.
}
