% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{default_amis_params}
\alias{default_amis_params}
\title{Produce list containing the default AMIS parameters}
\usage{
default_amis_params()
}
\value{
list containing the default AMIS parameters
}
\description{
Produce list containing the default AMIS parameters
}
