% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{evaluate_likelihood}
\alias{evaluate_likelihood}
\title{Evaluate likelihood for each additional simulation for a single timepoint}
\usage{
evaluate_likelihood(prevalence_map, prev_sim, amis_params)
}
\arguments{
\item{prevalence_map}{A list containing objects \code{data} (an L x M matrix of data);
and \code{likelihood} a function taking arguments \code{data} (a matrix of data as above),
\code{prevalence} (a matrix of output from the transmission model) and optional logical \code{log}, which returns the vector of (log)-likelihoods.}

\item{prev_sim}{A vector of simulated prevalences}

\item{amis_params}{A list of parameters, e.g. from \code{\link{default_amis_params}}.}
}
\value{
A locations x simulations matrix of (log-)likelihoods.
}
\description{
Implements analytical likelihoods if a likelihood function is available; otherwise histogram or empirical
likelihoods are generated based on samples from a geostatistical map.
}
