% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{evaluate_mixture}
\alias{evaluate_mixture}
\title{Evaluate the mixture function as returned by \link{mclustMix}}
\usage{
evaluate_mixture(parameters, nsamples, weight_matrix, mixture)
}
\arguments{
\item{parameters}{A matrix m x 2 matrix containing the sampled values for the
2 parameters.}

\item{nsamples}{The size of the parameter subset to evaluate the mixture
function from (integer)}

\item{weight_matrix}{The weight_matrix as returned by \link{compute_weight_matrix}}

\item{mixture}{The mixture function as returned by function \link{mclustMix}}
}
\value{
A list of the mixture components (see function \code{\link{mclustMix}})
\describe{
\item{\code{alpha}}{The probability?}
\item{\code{muHat}}{The mean?}
\item{\code{SigmaHat}}{The variance?}
\item{\code{G}}{??}
\item{\code{cluster}}{??}
}
}
\description{
Evaluate the mixture function MIXTURE for an subset of NSAMPLES parameters
values sampled according to the weight matrix WEIGHT_MATRIX.
}
\seealso{
\code{\link{mclustMix}}
}
