% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{fit_mixture}
\alias{fit_mixture}
\title{Fit an MVN mixture model using mclust}
\usage{
fit_mixture(dat, max.components = 10)
}
\arguments{
\item{dat}{a MATRIX or dataframe containing the observations to cluster.}

\item{max.components}{A postive integer specifying the maximum number of components to fit in the mixture.}
}
\value{
list containing selected output from \code{mclust}:
\describe{
\item{\code{G}}{the best number of components G.}
\item{\code{probs}}{vector of cluster probabilities (mixing weights).}
\item{\code{Mean}}{matrix of cluster means.}
\item{\code{Sigma}}{array of cluster covariance matrices.}
\item{\code{BIC}}{The BIC of the chosen mixture.}
\item{\code{ModelName}}{The model name from the package \code{mclust}.}
}
}
\description{
Can handle 1D as well as multivariate clustering. \code{dat} must have \code{nrow} observation and \code{ncol}
dimensions, even if the number of dimensions is 1.
Uses BIC to determine the best number of components, up to max.components.
}
