% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{get_initial_parameters}
\alias{get_initial_parameters}
\title{Sample initial parameters from prior}
\usage{
get_initial_parameters(n)
}
\arguments{
\item{n}{The number of samples to draw for each parameter.}

\item{min}{The lower bound for the uniform distribution.}

\item{max}{The upper bound for the uniform distribution.}
}
\value{
A Nx2 matrix
}
\description{
Sample N initial values for two parameters, uniformly distributed
between MIN and MAX.
}
\examples{
get_initial_parameters(100, 0.05, 0.175)
}
