% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{sample_new_parameters}
\alias{sample_new_parameters}
\title{Sample new parameters}
\usage{
sample_new_parameters(mixture, n_samples, df, prior, log)
}
\arguments{
\item{mixture}{A list of mixture components as returned by
\code{\link{evaluate_mixture}}}

\item{n_samples}{A number of new parameters to sample (integer)}

\item{df}{The degrees of freedom for the t-dsitributed proposal distribution.}

\item{prior}{list containing the functions \code{rprior} and \code{dprior}}

\item{log}{A logical indicating if densities}
}
\value{
A list containing \code{params}, an \code{n_samples} x d matrix containing the sampled parameter values and
\code{prior_density}, the corresponding vector of prior densities.
}
\description{
This function generates \code{nsamples} new model parameter values according the
t-distribution with \code{df} degrees of freedom and the mixture components \code{mixture}.
}
\seealso{
\code{\link{fit_mixture}}
}
