% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{weighted_mixture}
\alias{weighted_mixture}
\title{Fit mixture to weighted sample}
\usage{
weighted_mixture(parameters, nsamples, weights, log = F)
}
\arguments{
\item{parameters}{An N x d matrix containing the sampled values for the d parameters.}

\item{nsamples}{The number of parameter to resample as data to fit the mixture to.}

\item{weights}{A vector of weights with length N.}

\item{log}{logical indicating if weights are logged.}
}
\value{
A list of the mixture components (see function \code{\link{fit_mixture}})
\describe{
\item{\code{probs}}{The mixture weights}
\item{\code{Mean}}{The means of the components}
\item{\code{Sigma}}{The covariance matrices of the components}
\item{\code{G}}{Number of components}
\item{\code{BIC}}{BIC of fitted mixture}
\item{\code{ModelName}}{Model name from package mclust}
}
}
\description{
Weights are implemented by using systematic resampling to obtain an unweighted set of parameters.
An unweighted mixture is then fitted using \code{fit_mixture}.
}
\seealso{
\code{\link{fit_mixture}}
}
