% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{write_model_input}
\alias{write_model_input}
\title{Write input file required by transmission model}
\usage{
write_model_input(seeds, beta, input_file)
}
\arguments{
\item{seeds}{A vector containing seed value for each parameter sample
(double)}

\item{beta}{A vector containing samples of the beta parameter (double)}

\item{input_file}{The name of the input file (string)}
}
\description{
Write a 2 columns csv file named INPUT_FILE to disk. The first column SEEDS
contains the seed for each parameter value. Second column BETA contains
corresponding parameter values.
}
