% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS.R
\name{amis}
\alias{amis}
\title{Run the AMIS algorithm to fit the transmission model to the map}
\usage{
amis(prevalence_map, transmission_model, prior, amis_params, seed = NULL)
}
\arguments{
\item{prevalence_map}{An L x M matrix containing samples from the prevalence map, where L is the number of locations and M the number of samples.
The location names are inherited from \code{rownames(prevalence_map)} if possible.}

\item{transmission_model}{A function taking a vector of seeds and a matrix of parameter vectors as inputs.}

\item{prior}{A list containing the functions \code{dprior} and \code{rprior} (density and RNG).
\code{rprior} must produce an n by d MATRIX of parameters, even when d=1.
parameter names are inherited from the \code{colnames} from the output of \code{rprior} if possible.}

\item{amis_params}{A list containing the control parameters for the AMIS algorithm
\describe{
\item{\code{delta}the smoothing parameter in the empirical RN derivative (usually 0.01).}
\item{\code{nsamples}the number of new samples drawn within each AMIS iteration.}
\item{\code{mixture_samples}the number of samples used to represent the weighted parameters in the mixture fitting.}
\item{\code{df}the degrees of freedom in the t-distributions, used to yield a heavy tailed proposal.}
\item{\code{target_ess}the target effective sample size.}
\item{\code{log} logicalindicating if calculations are to be performed on log scale.}
\item{\code{max_iters}maximum number of AMIS iterations.}
}}

\item{seed}{Optional seed for the random number generator}
}
\value{
A list containing a dataframe of the sampled parameters, simulation seed, and weight in each location, plus a vector
#' called ess containing the obtained ess at each location.
}
\description{
Run the AMIS algorithm to fit the transmission model to the map
}
