% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{calculate_ess}
\alias{calculate_ess}
\title{Compute the current effective sample size}
\usage{
calculate_ess(weight_mat)
}
\arguments{
\item{weight_mat}{The weight matrix. A n x m matrix with n the number of IUS
and m the number of sampled parameter values.}
}
\value{
A vector containing the ESS value for each IU.
}
\description{
This function returns the effective sample size (ESS) for each Implementation
Unit (IU). For each column of the weight matrix WEIGHT_MAT, the ESS component
is computed as
}
\details{
\deqn{\left(\sum_{i=1}^{N} w_{i}^2 \right)^{-1}},

where \eqn{$N$} is the number of sampled parameter values for each IU.
}
\seealso{
\code{\link{compute_weight_matrix}}
}
