% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{compute_prior_proposal_ratio}
\alias{compute_prior_proposal_ratio}
\title{Compute the prior/proposal ratio}
\usage{
compute_prior_proposal_ratio(components, param, dprop)
}
\arguments{
\item{components}{A list of mixture components made of
\describe{
\item{\code{GG}}{A numeric vector}
\item{\code{Sigma}}{A list}
\item{\code{Mean}}{A list}
\item{\code{PP}}{A list}
}}

\item{param}{A 2 column matrix containing the values for the two parameters.}

\item{dprop}{The proposal distribution as returned by \code{mvtComp()$d}.}
}
\value{
A vector contaning the prior/proposal ratio for each row in
\code{param}
}
\description{
This function returns the ratio between the prior and proposal distribution
for each sampled parameter value (i.e. each row in PARAM).
}
\details{
This function returns
\deqn(\frac{\pi(\theta \emph{j)}{\frac{1}{N_1 + ... + N_t}\sum}{u=1}^{t}N_u \phi_u(\theta _j)})
See step (4) if the AMIS algorithm in

Integrating Geostatistical Maps And Transmission Models Using Adaptive
Multiple Importance Sampling
Renata Retkute, Panayiota Touloupou, Maria-Gloria Basanez,
T. Deirdre Hollingsworth, Simon E.F. Spencer
medRxiv 2020.08.03.20146241;
doi: \url{https://doi.org/10.1101/2020.08.03.20146241}
}
