% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{compute_weight_matrix}
\alias{compute_weight_matrix}
\title{Compute weight matrix}
\usage{
compute_weight_matrix(prev_data, prev_sim, delta, first_weight)
}
\arguments{
\item{prev_data}{The geostatistical prevalence data. A n x m matrix where n
is the number of IUs and m the number of prevalence samples. (double)}

\item{prev_sim}{A vector containing the siumulated prevalence value for each
parameter sample. (double)}

\item{delta}{A number. (double)}

\item{first_weight}{A vector containing the values for the right hand side of
the weight expression. Should be of the same length as \code{prev_sim}.}
}
\description{
Compute matrix describing the weights for each parameter sampled, for each
Implementation Unit (IU). One row per sample, one column per IU.  Each weight
is computed based on the Randon-Nikodim derivative, taking into account
geostatistical prevalence data for the specific IU and the prevalence value
computed from the transmission model for the specific parameter sample.
}
