% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{sample_new_parameters}
\alias{sample_new_parameters}
\title{Sample new parameters}
\usage{
sample_new_parameters(clustMix, n_samples, rprop)
}
\arguments{
\item{clustMix}{A list of mixture components as returned by
\code{\link{evaluate_mixture}}}

\item{rprop}{The proposal distribution as returned by \code{mvtComp()$r}.}

\item{nsamples}{A number of new parameters to sample (integer)}
}
\value{
A NSAMPLES x 2 matrix containing the sampled parameter values.
}
\description{
This function generates NSAMPLES new model parameter values according the
proposal distribution RPROP and the mixture components CLUSTMIX.
}
\seealso{
\code{\link{mvtComp}}
}
