% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{update_according_to_ess_value}
\alias{update_according_to_ess_value}
\title{Update weight matrix}
\usage{
update_according_to_ess_value(weight_matrix, ess, target_size)
}
\arguments{
\item{weight_matrix}{The weight_matrix as returned by
\link{compute_weight_matrix}}

\item{ess}{The effective sample size vector as returned by
\link{calculate_ess}}

\item{target_size}{A number representing the target size for the sample.}
}
\value{
The updated weight matrix. Its size is unchanged.
}
\description{
This function sets to 0 the rows of the weight matrix WEIGHT_MATRIX for which
the effective sample size ESS is above a target size TARGET_SIZE.
}
