% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMIS_steps.R
\name{update_mixture_components}
\alias{update_mixture_components}
\title{Update the components of the mixture}
\usage{
update_mixture_components(clustMix, components, t)
}
\arguments{
\item{clustMix}{A list of mixture components as returned by
\code{\link{evaluate_mixure}}}

\item{components}{A list of mixture components made of
\describe{
\item{\code{GG}}{A numeric vector}
\item{\code{Sigma}}{A list}
\item{\code{Mean}}{A list}
\item{\code{PP}}{A list}
}}

\item{t}{The current iteration index (integer)}
}
\value{
The updated \code{components} list
}
\description{
This function updates the components of the mixture COMPONENTS according to
the current mixture CLUSTMIX generated at iteration T.
}
\seealso{
\code{\link{evaluate_mixture}}, \code{\link{mclustMix}}
}
